[HUBOVRVW-760] Fix initialization of widgets
Since marko-widgets@6.3.0 the state of a widget is no longer stored in
the html nodes but globally in the variable
`window.$markoWidgetsState`. Because of this widgets with a
getInitialState method no longer had a state during tests.

This PR fixes the behavior by populating the global state variable
before initializing the widget.